+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
+
+## [0.5.0] - 2020-12-20
 ### Added
 - `FaceInfo::post_script_name`
 - `FaceInfo::monospaced`
@@ -25,7 +27,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ### Fixed
 - Stretch processing. `ttf-parser` was incorrectly parsing this property.
 
-[Unreleased]: https://github.com/RazrFalcon/fontdb/compare/v0.4.0...HEAD
+[Unreleased]: https://github.com/RazrFalcon/fontdb/compare/v0.5.0...HEAD
+[0.5.0]: https://github.com/RazrFalcon/fontdb/compare/v0.4.0...v0.5.0
 [0.4.0]: https://github.com/RazrFalcon/fontdb/compare/v0.3.0...v0.4.0
 [0.3.0]: https://github.com/RazrFalcon/fontdb/compare/v0.2.0...v0.3.0
 [0.2.0]: https://github.com/RazrFalcon/fontdb/compare/v0.1.0...v0.2.0